Agentic_RAG/rag_qwen_bge_compare/requirements.txt
torch>=2.1.0
transformers>=4.41.0
accelerate>=0.31.0
# bitsandbytes==0.43.1
sentence-transformers>=2.7.0
chromadb>=0.5.5
numpy
pandas
tqdm
pyyaml
